Get MTU, start to implement write proc

+++ dfu.h
@@ -12,6 +12,6 @@ bool dfu_create_object(uint8_t type, uint32_t size);
 bool dfu_object_write(FILE* fp);
 bool dfu_get_crc(void);
 bool dfu_object_execute(void);
-bool dfu_object_write_procedure(uint32_t type, FILE* fp);
+bool dfu_object_write_procedure(uint8_t type, FILE* fp);
 
 #endif
